refactor: move UiRoot and its spawning system to ui module for now
