+++ src/control_display.rs
@@ -1,9 +1,11 @@
 //! Plugin providing functionality for the gameplay UI element showing the current controls.
 use crate::{
     camera::PlayZonePortion,
-    gameplay::components::UiRoot,
     movement_table::{Direction, MovementTable, DIRECTION_ORDER},
-    ui::font_scale::{FontScale, FontSize},
+    ui::{
+        font_scale::{FontScale, FontSize},
+        UiRoot,
+    },
     GameState,
 };
 use bevy::prelude::*;
