--- src/gameplay/components.rs
@@ -1,4 +1,0 @@
-use bevy::prelude::*;
-
-#[derive(Copy, Clone, Eq, PartialEq, Debug, Default, Hash, Component)]
-pub struct UiRoot;
