+++ src/ui/mod.rs
@@ -24,6 +24,7 @@ pub struct UiPlugin;
 impl Plugin for UiPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugin(PreviousComponentPlugin::<Interaction>::default())
+            .add_startup_system(spawn_ui_root)
             .add_event::<actions::UiAction>()
             .add_plugin(font_scale::FontScalePlugin)
             .add_system(text_button::text_button_visuals.run_not_in_state(GameState::AssetLoading))
@@ -40,4 +41,24 @@ impl Plugin for UiPlugin {
                 .add_system(actions::debug_print_action);
         }
     }
+}
+
+// TODO: delete this in favor of z-indices after upgrading to bevy 0.9
+#[derive(Copy, Clone, Eq, PartialEq, Debug, Default, Hash, Component)]
+pub struct UiRoot;
+
+fn spawn_ui_root(mut commands: Commands) {
+    commands
+        .spawn_bundle(NodeBundle {
+            color: UiColor(Color::NONE),
+            style: Style {
+                size: Size {
+                    width: Val::Percent(100.),
+                    height: Val::Percent(100.),
+                },
+                ..Default::default()
+            },
+            ..Default::default()
+        })
+        .insert(UiRoot);
 }
